+++ controllers/eventsource_controller_test.go
@@ -42,7 +42,7 @@ var _ = Describe("Controllers: eventSource controller", func() {
 		watcherCtx, cancel = context.WithCancel(context.Background())
 		evaluation.InitializeManager(watcherCtx, textlogger.NewLogger(textlogger.NewConfig(textlogger.Verbosity(1))),
 			testEnv.Config, testEnv.Client, randomString(), randomString(), randomString(),
-			libsveltosv1beta1.ClusterTypeCapi, 3, false)
+			libsveltosv1beta1.ClusterTypeCapi, int64(3), false)
 	})
 
 	AfterEach(func() {
